Simplifies ReactTextInputShadowNode base class.
